requirements.txt
antiorm==1.2.1
argcomplete==1.8.2
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.5.3
cachetools==4.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cssselect==1.0.3
db==0.1.1
docx2txt==0.8
EbookLib==0.15
extract-msg==0.23.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
geograpy3==0.1.2
google-api-core==1.16.0
google-auth==1.10.1
google-cloud-speech==1.3.1
googleapis-common-protos==1.51.0
grpcio==1.26.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
IMAPClient==2.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
jellyfish==0.7.2
jieba3k==0.35.1
Jinja2==2.10.1
lxml==4.4.0
MarkupSafe==1.1.1
more-itertools==7.2.0
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.0
olefile==0.46
opencv-python==4.1.2.30
packaging==19.1
pandas==0.25.0
pdfminer.six==20181108
pdfminer3k==1.3.1
Pillow>=6.2.0
pluggy==0.12.0
ply==3.11
protobuf==3.11.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.7.15
pycryptodome==3.8.2
pyparsing==2.4.2
PyPDF3==1.0.1
pytest==5.0.1
python-dateutil==2.8.0
python-pptx==0.6.5
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
rsa==4.0
six==1.10.0
slate3k==0.5.3
sortedcontainers==2.1.0
soupsieve==1.9.2
SpeechRecognition==3.6.3
tinysegmenter==0.3
tldextract==2.2.1
tqdm==4.32.2
tzlocal==1.5.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
xlrd==1.0.0
XlsxWriter==1.1.8
zipp==0.5.2
